// src/app/(root)/forum-prestacao-contas/interface.ts
export const mandatoSchema = z.object({
  data_inicio: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, "Formato inválido (YYYY-MM-DD)"),
  data_fim: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, "Formato inválido (YYYY-MM-DD)"),
  ctgom_id: z.number().min(1, "CTGOM é obrigatório"),
  membros: z.array(z.object({
    nome: z.string().min(1, "Nome é obrigatório"),
    funcao: z.string().min(1, "Função é obrigatória"),
    contacto: z.string().min(1, "Contacto é obrigatório"),
    email: z.email("Email é obrigatório"),
  })),
  uploads: z.array(z.object({
    nome: z.string().min(1, "Título é obrigatório"),
    tipo: z.string().min(1, "Descrição é obrigatória"),
  })).optional(),
 })
 export type MandatoFormValues = z.infer<typeof mandatoSchema>